refactor(viewer): move fetchActivities
- Moves the following functions from `routes` to `utils`:
    - `fetchActivities`
    - `fetchMetaActivities`
    - `loadActivity`

issue codemonger-io/mumble#12
